Add ability for clients to print which engine they are connected to.
When working with a lot of engine instances and clients that may end up
connecting to any one of them it can be extremely useful to know which
engine a given client connected to.

The engine can be configured w/ a name via a `DAGGER_ENGINE_NAME` env,
otherwise it will just default to the hostname. Clients will now log
that when they connect to it.

Signed-off-by: Erik Sipsma <erik@dagger.io>