dagger run: don't attach stdin
Some applications detect the presence of a TTY on stdin/stdout/stderr
and use that to determine whether to render a TUI of their own. In that
case the application will likely just hang.

Also: this was never guaranteed to work as expected anyway, because both
the Dagger TUI and the application will be competing for stdin's
attention.

Signed-off-by: Alex Suraci <alex@dagger.io>