fix out-of-order event delivery (#4996)
This one's pretty straightforward. Previously we would concurrently
write to all event readers, but that meant rapid writes to the same
reader could be delivered out-of-order. This was especially likely to
happen to vertexes that quickly start and finish.

It looks like the goal was to prevent one slow consumer from blocking
all the rest, but I think we're fine; the TUI has an infinitely buffered
pipe, and the telemetry is also pushed into a queue and sent in batches,
so in practice it should be OK to just block.

Signed-off-by: Alex Suraci <alex@dagger.io>