Use background context to release gateway containers.
In the code for obtaining stdout/stderr of a failed exec, we were
passing the context to the deferred Release of the gateway container
that is used for all the other calls. This context had a failsafe 30
second timeout so that obtaining stdout/stderr didn't inadvertently
block for excessive amounts of time.

If that context ended up cancelled, the Release call would not actually
happen, which meant that all the cache refs for the created container
would continue to be held open in the engine even after the container
fully exited.

We also were previously including all mounts in this container,
including any cache mounts. If the cache mount was mode LOCKED, it was
entirely possible for that timeout to be hit while the container waited
for the cache mount to become available.

The end result was that if a locked cache mount was included with a
failed exec, it was possible to leak the reference to that cache mount
in the engine, making it unavailable to any other exec trying to use it
from any other client. It would only be released once the session
shutdown.

The change here fixes Release to just use a background context. It also
updates the gateway container to not include cache mounts or tmpfs
mounts since they can't contain stdout/stderr even when those are
redirected.

Unforunately, this behavior is hard to reliably test since it relies on
a parallel exec obtaining the locked cache mount after we have created
one of these gateway containers but before Release gets called on it.

This bug was fairly reproducible when manually running some load tests
though, and I have no been able to reproduce it anymore after this
change.

Signed-off-by: Erik Sipsma <erik@dagger.io>