redo atomic host file updates
this wasn't really atomic before: it would _rename_ the live file, and
then start writing to it. if dnsmasq itself tried to read the host file,
it could either see that the file doesn't exist, or see a partially
written file.

now we'll write to a new path and then rename it over the live one,
which is atomic.

Signed-off-by: Alex Suraci <alex@dagger.io>