cache: Use URLs to sync layers and mounts. (#5050)
* cache: Use URLs to sync layers and mounts.

This replaces the need for the engine to use the S3 API directly and
have credentials for accessing the bucket. There's a few benefits:
1. Engine doesn't need to be setup with cross-account access to Dagger's
   S3 buckets, which greatly simplifies configuration for users and
   gives Dagger flexibility to modify buckets without users having to
   make updates.
2. It's more secure as the pre-signed URLs are tightly scoped in terms
   of capabilities and expire after a timeout.
3. It greatly reduces the amount of S3-specific code in the engine,
   which will make it much easier to generalize to other storage
   backends in the future.

This change required a modification to how cache mounts are packed.
Previously, we were using rclone inside a dagger exec to sync these. It
was planned to eventually optimize this more, but the implementation was
still just uploading/downloading each file in the cache mount
individually and uncompressed, which didn't scale well to cache mounts
with large numbers of files and/or disk size.

This also did not work with the pre-signed URL approach as it would
require getting a pre-signed URL for each individual file. So we now
remove rclone and instead just directly pack/unpack cache mounts using
the same containerd libraries used to pack/unpack layers. Each cache
mount is essentially treated as a single large layer.

rclone is worth considering a revisit in the future as it could have
better performance with more tuning+customization on our part, plus
unique features like lazy-loading via FUSE mounts. But for now, this new
approach is more efficient in addition to enabling exclusive use of
pre-signed URLs.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Remove now unused OperatorClient code.

This was allowing the engine to run execs for populating cache mounts,
but now that code has been refactored to read/write cache mounts
directly.

It's likely we'll want this again in the future, but can just retrieve
it from git history at that time.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Ignore false-positive linter complaint about http body close.

AFAICT the body is closed in all codepaths, so unless I'm missing
something this isn't relevent and can be ignored.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* appease linter

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Ensure synced cache mount is emptied on sync failure.

Signed-off-by: Erik Sipsma <erik@dagger.io>

---------

Signed-off-by: Erik Sipsma <erik@dagger.io>