cache: Support extra headers for pre-signed URLs (#5068)
For certain settings, S3 pre-signed URLs will sign the URL with the
setting name and value, but doesn't include it in the URL as a query
string. It's expected that the user of the URL include it as a header
with the same value when actually making the request.

Previously, this was just Content-Length, which we handled as a special
case but it turns out another such header is `x-amz-acl`, which we
currently need to use in order to ensure that uploading to a
cross-account bucket will work. This only due to the fact that the cache
service still runs on-prem; once it's in the cloud then the generator of
the pre-signed URL will always be the owner of the bucket. But until
that time, this is required.

Implemented this by having the cache service just give the engine a set
of extra headers to include with its upload requests. This keeps the
engine code S3-agnostic.

Signed-off-by: Erik Sipsma <erik@dagger.io>