Truncate stdout/stderr from shim rather than session.
This is primarily needed to workaround an issue in Buildkit where a
series of bugs can result in container stdout/stderr resulting in runc
getting blocked trying to write to its output pipes and never exit.

That needs an upstream fix but we can make it much rarer by reducing the
amount of output our shim can send in the case where its printing
stdout/stderr after an exec failure, which is the main place we hit the
bug previously.

We were already truncating that output anyways, but we were doing it in
the session. This updates the truncation to happen in the shim itself,
which is slightly more complicated but avoids the bug and is technically
more performant in general anyways.

Signed-off-by: Erik Sipsma <erik@dagger.io>