Configure Dependabot to bump dependencies monthly
"daily" results in too much noise. It makes it more difficult to spot
"real" contributions vs. "noisy ones".

With a reduced frequency (monthly instead of daily), this gives us the
chance to update the things that we care about ourselves. It is known
for some updates to result in breakages further down the pipeline, e.g.
https://github.com/dagger/dagger/pull/4850

FTR:
https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuring-dependabot-version-updates

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>