tooling: Reduce frequency of Dependabot PRs (#5099)
* Configure Dependabot to bump dependencies monthly

"daily" results in too much noise. It makes it more difficult to spot
"real" contributions vs. "noisy ones".

With a reduced frequency (monthly instead of daily), this gives us the
chance to update the things that we care about ourselves. It is known
for some updates to result in breakages further down the pipeline, e.g.
https://github.com/dagger/dagger/pull/4850

FTR:
https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuring-dependabot-version-updates

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Append extra labels to Dependabot PRs

We were missing a few dependency update areas, e.g. engine, cli,
tooling, etc.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Configure Dependabot to ignore patch updates

FTR:
https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#ignore

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Configure Dependabot to bump Go deps in /internal/mage

Before this, I could not find a single Dependabot update of
/internal/mage/go.mod

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

---------

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>