cache: Reduce noise by removing unreferenced keys.
Before this change, engines running for a long time were starting to
output tons of debug logs about skipping cache results due to them not
being found. In one case, of the 1,048,576 most recent log lines for an
engine, 1,045,160 of them were those skipping cache result lines...

This log message is not actually indicative of anything wrong; it's
expected to happen when caches have been pruned or if the cache ref is
lazy (i.e. an image that has not been pulled locally). Typically, the
cache keys associated with these missing results are deleted:
1. When buildkitd starts
2. When a client makes an explicit Prune call with the Control API
3. When the `Results` method of solver.CacheManager is called with a
   cache key whose results end up not being found.

In our case, it's entirely possible to not hit any of those conditions
for a very long time, which meant that these unreferenced cache keys
could accumulate and result in those enormous number of log lines being
printed every time an export ran (currently once a minute).

The fix here calls Release on the cache result if we hit the case of it
not being found, which means the underlying cache result has been
pruned.

This fixes the log noise and also probably improves performance due to
not having to spend time re-walking keys that have missing results.

Signed-off-by: Erik Sipsma <erik@dagger.io>