cache: deduplicate layer pushes in single update
Before this, we would skip pushing the same layer multiple times across
different updates due to the fact that the cache service wouldn't ask
for engines to push a layer it already has.

However, the same layer could end up being in the same update batch many
times because it can appear across all the refs being exported an
arbitrary number of times. Now, the engine keeps track of what layers it
has pushed so far in this update in order to skip what it already took
care of.

Signed-off-by: Erik Sipsma <erik@dagger.io>