Grab bag of minor logging+performance improvements in cache (#5116)
* cache: Reduce noise by removing unreferenced keys.

Before this change, engines running for a long time were starting to
output tons of debug logs about skipping cache results due to them not
being found. In one case, of the 1,048,576 most recent log lines for an
engine, 1,045,160 of them were those skipping cache result lines...

This log message is not actually indicative of anything wrong; it's
expected to happen when caches have been pruned or if the cache ref is
lazy (i.e. an image that has not been pulled locally). Typically, the
cache keys associated with these missing results are deleted:
1. When buildkitd starts
2. When a client makes an explicit Prune call with the Control API
3. When the `Results` method of solver.CacheManager is called with a
   cache key whose results end up not being found.

In our case, it's entirely possible to not hit any of those conditions
for a very long time, which meant that these unreferenced cache keys
could accumulate and result in those enormous number of log lines being
printed every time an export ran (currently once a minute).

The fix here calls Release on the cache result if we hit the case of it
not being found, which means the underlying cache result has been
pruned.

This fixes the log noise and also probably improves performance due to
not having to spend time re-walking keys that have missing results.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* cache: Add ReleaseUnreferenced method secretely called by buildkit.

Buildkit will call this method on a solver.CacheManager instance in the
control API's Prune method, but only by doing a type conversion to an
anonymous interface and then calling that...

We don't have an official mechanism for calling Prune right now, but it
is possible via `buildctl` inside the engine container, so may as well
add this method to our cache manager implementation too.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* cache: add extra debug logs

Signed-off-by: Erik Sipsma <erik@dagger.io>

* cache: deduplicate layer pushes in single update

Before this, we would skip pushing the same layer multiple times across
different updates due to the fact that the cache service wouldn't ask
for engines to push a layer it already has.

However, the same layer could end up being in the same update batch many
times because it can appear across all the refs being exported an
arbitrary number of times. Now, the engine keeps track of what layers it
has pushed so far in this update in order to skip what it already took
care of.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* cache: use ticker for import/export timing

Signed-off-by: Erik Sipsma <erik@dagger.io>

---------

Signed-off-by: Erik Sipsma <erik@dagger.io>