go sdk: Add ability to explicitly set engine conn.
This is technically a special ability that's only available to the Go
SDK, but I can't think of any harm in granting it. It's very convenient
for our own internal uses cases where we want to use the Go SDK in parts
of the engine code like the CLI and/or tests.

Signed-off-by: Erik Sipsma <erik@dagger.io>