Refactor project API; remove old stuff, add commands.
This migrates the ancient project code into the modern world, it now
follows the same conventions as the rest of our APIs.

It also strips down the API to just what we care about right now, which
are commands.

Some of the deleted code here will likely need resurrection in the
near-ish future (i.e. typescript runtime, support for recursively loading
dependencies, etc.), but it's easier to start from a clean base and
re-add that as needed.

Signed-off-by: Erik Sipsma <erik@dagger.io>