Project Commands (#5060)
* first pass entrypoints

Signed-off-by: Erik Sipsma <erik@dagger.io>

* just barely working python entrypoint

Signed-off-by: Erik Sipsma <erik@dagger.io>

* tweak messages to help awkward silence when command is uncached

Signed-off-by: Erik Sipsma <erik@dagger.io>

* add descriptions to python entrypoint example that show up in --help

Signed-off-by: Erik Sipsma <erik@dagger.io>

* wip CI event integration

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Add dagger.Context to give go entrypoint pre-made client.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Migrate basic linting steps from mage->entrypoints.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* fixups made while updating description

Signed-off-by: Erik Sipsma <erik@dagger.io>

* go codegen: handle both optional and non-null lists of objects

Signed-off-by: Erik Sipsma <erik@dagger.io>

* go sdk: Add ability to explicitly set engine conn.

This is technically a special ability that's only available to the Go
SDK, but I can't think of any harm in granting it. It's very convenient
for our own internal uses cases where we want to use the Go SDK in parts
of the engine code like the CLI and/or tests.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Refactor project API; remove old stuff, add commands.

This migrates the ancient project code into the modern world, it now
follows the same conventions as the rest of our APIs.

It also strips down the API to just what we care about right now, which
are commands.

Some of the deleted code here will likely need resurrection in the
near-ish future (i.e. typescript runtime, support for recursively loading
dependencies, etc.), but it's easier to start from a clean base and
re-add that as needed.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Remove ConfigPath and NoExtensions from engine/sdk configs.

The new approach just has callers load projects, so passing those config
options around everywhere was unneeded and confusing.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Cleanup project dir + dagger.json path config.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* catch other sdks up on latest schema

Signed-off-by: Erik Sipsma <erik@dagger.io>

* remove old test

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Formalize lint-daggerized.yml a bit more

Signed-off-by: Erik Sipsma <erik@dagger.io>

* appease linter

Signed-off-by: Erik Sipsma <erik@dagger.io>

* re-enable cloud token in workflows

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Switch to specifying commands with ":" as separator.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* misc cleanups

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Add comments to project state fields.

Signed-off-by: Erik Sipsma <erik@dagger.io>

---------

Signed-off-by: Erik Sipsma <erik@dagger.io>