Scrub secret plaintext arg from progrock vertex.
Problem before was that we'd serialize all the graphql args and values
for every resolver, which included the secret plaintext value.

The fix here is an attempt to give a slightly more general+reusable
solution than just putting in a hardcoded `if` in the progrock related
code to check for this fieldname+argname pair. It makes the type of
the plaintext in our code to be `type SecretPlaintext string` and
gives that a `MarshalText` method so progrock shows that instead of the
actual value.

Signed-off-by: Erik Sipsma <erik@dagger.io>