tui: Add fallback for runtime dir permission denied errors.
It seems it's common in dev environments running in containers for the
user runtime dir to not exist and not be creatable w/out sudo, which
results in permission denied errors when the tui sets up the progrock
socket.

This adds a fallback to use the home cache dir in such cases. This isn't
as ideal since that dir isn't a tmpfs, but is better than failing I think.
A warning is printed when this happens to let users know.

Signed-off-by: Erik Sipsma <erik@dagger.io>