bump progrock for gRPC + resilient draining
Previously the RPC flow required clients to explicitly send a Close RPC
call, which wouldn't be sent for services because they are kill -9'd.

Now the RPC flow uses gRPC under the hood and simply uses a graceful
shutdown to wait for any in-flight RPC requests to finish, which
accomplishes the same thing while also allowing RPC clients to suddenly
disappear.

Signed-off-by: Alex Suraci <alex@dagger.io>