avoid mutation in overrideProgress
Fixes a panic caused by calling Container.ID or Container.Digest while
the container is being started as a service.

We don't really want to mutate in this scenario anyway; it's purely a
runtime concern, and should not affect the original Container.

Ultimately this whole function can go away once we're fully switched to
Progrock.

Signed-off-by: Alex Suraci <alex@dagger.io>