ci: Fix lint-daggerized.yml error
Found from
https://github.com/dagger/dagger/actions/runs/5147441125. The `uses`
must have `./` to point to local action. And fixes 2 issues in the
workflow:

First, fix the daggerized job to use `${{ github.ref }}` instead of
`${{ github.sha }}` when setting `DAGGER_PROJECT` because when using
sha, the dagger didn't know the sha commit when sending a Pull Request
from another repo.

Second, fix the `dagger do` argument and passing `--repo` and `--branch`
to the project.

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>