Add "root" to dagger.json; remove --config flag.
This simplifies dagger project and dagger do by getting rid of the
--config flag and only asking for a --project flag.

The previous need for the two was that project specified the "root"
which needed to be loaded in order for the project to run, whereas
config specified the path to the dagger.json.

This is inherently confusing, but gets even more confusing when you
start needing to deal with corner cases around whether the config path
should be considered relative to the caller's workdir or relative to the
project dir.

Now, we have a new "root" field in dagger.json that, if set, will be
loaded as the root context for the project. This allows us to get rid of
the config flag and just use --project to point to dagger.json (or its
parent dir).

For git projects, we now consolidate all settings, including the newly
added "subdir" as query parameters on the "git://" URI. Query params are
not beautiful, but it's probably an upgrade from the previous use of
fragments, which break shells due to interpretation of "#" as a comment.

Signed-off-by: Erik Sipsma <erik@dagger.io>