run services via gateway interface instead
There were a few tricky parts to this.

Problem A: It doesn't seem possible to take an arbitrary Container and
run it via the Gateway instead. The Container already "bakes in" an LLB
definition that refers to the _result_ of running the service. The only
thing you can really do is try to Solve() it.

Solution A: change the API such that Services are no longer just
Containers. Instead, you call Container.Service which takes the exact
same params as Container.WithExec, only it returns a Service instead of
a Container.

Problem B: it's harder to generate service hostnames. Previously we
could just take the digest that WithExec baked in, but now that we're
not doing that, we have to come up with our own content-addressing
scheme. The naive thing to try is taking a digest of the Service, akin
to a ContainerID, ContainerIDs aren't stable.

Solution B: I implemented a helper to take a stable digest of an
arbitrary struct. This works by recursing through structs, maps, slices,
etc. and digesting types, with a special-case for *pb.Definition which
was the source of unstable hashing as its inner Def field is
topographically ordered.

Problem C: bizarre behavior with running the same service across Dagger
sessions (i.e. parallel test runs), causing each session to start the
same service + hostname, and some of them to fail when resolving the
host to an IP.

Solution C: dedupe services within a session, but not across sessions.
Maintain a global map of all services, keeping track of their
starting/running state to handle parallelism, and injecting a "session
ID" into each Service so that equivalent containers don't share
hostnames across Dagger sessions.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>