Send Progrock updates to Cloud (#5297)
* send progrock events to Cloud

This is an incremental step that maintains compatibility with the Dagger
Cloud API.

Groups are mapped to pipeline paths. (TODO: pipeline descriptions are
currently droped.)

Vertex <-> Group membership is handled client-side; an Op will be
emitted for every group membership.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* preserve pipeline descriptions through progrock

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* remove erroneous [hide] handling

this was copy-pasta from Bass, not intended to be preserved in Dagger

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* clean up pre-Progrock progress support

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* bump progrock for minor UI fixes

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* bring back telemetry.Telemetry

this is used for engine logs - need to dedupe these, or maybe use one in
the other

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* support 'weak' pipelines

'weak' pipelines are ones that Dagger automatically adds, or a Buildkit
frontend adds via a ProgressGroup. They are treated differently from
user-specified pipelines.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* limit vertices to one group for now

Too difficult to figure out many-to-many vertex <-> groups in the
frontend. This is much simpler for now.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* JournalFile writes progrock updates

* remove JournalWriter (unused)
* remove internal/engine/journal/ (outdated)

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* convert otel-collector to progrock

* add telemetry.Pipeliner for tying vertices to pipeline.Paths

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* convert upload-journal to progrock

trivial now that there's telemetry.Writer

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* convert dagger-graph to progrock

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* use telemetry.Pipeliner as part of Writer

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* telemetry: consolidate Telemetry and Writer

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* skip pipeline console output tests for now

We're no longer putting this information into Buildkit, so the console
output no longer shows pipelines and such. To be clear, this format was
always a bit janky, often resulting in things like blank vertex names.
These tests can be brought back once Progrock has a proper console
output format.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* get rid of temporary 'meta' vertex

this was a temporary hack to show the Dagger Cloud URL in the TUI. will
figure this out in a separate PR.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* promote recorder completion to engine.Start

no need to tie this 100% to Buildkit events, in case we want to create
any other vertexes

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* use progrock console format, add --progress

* --silent no longer configures console output format, which was frankly
  confusing. --progress has been added instead.
* removed buildkit console format; we are now 100% Progrock!
* show Cloud URL and engine name in TUI

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* shim: don't use LogOutput during nesting

the Progrock socket should always be available, so this is pretty
pointless anyway

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* appease linter

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* remove pre-progrock code from Host.File

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* log console output from 'dagger session'

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* pipeline.Path: change to []*Pipeline

pretty tiny change that should reduce some memory pressure, since each
of these values never change. (this likely has very little actual
effect, just seems sensible.)

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* fix comment

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

---------

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>