project: partial support for returning host dirs/files.
This allows you to (usually) return a File or Directory from a command
that has a Local SourceOp somewhere in its LLB DAG. This is very
important as many projects will load source code from the project dir in
order to, e.g. build binaries/containers.

It works by using the session ID setting in local source ops, plus doing
a synchronous evaluate of the local op to ensure its cached.

Technically, it's still possible for the local ref to get pruned between
time of sync and time of use, but this solution is just meant to be
better than before, which always resulted in either an error or the
wrong dir getting loaded.

The long-term robust solution will come from better session sharing.

I considered an alternative solution of having the project API code
rewrite the LLB returned by projects to replace any local references
with the LLB for the project's exec op.

That approach has some nice advantages, but is such a headache-inducing
undertaking that I don't think it's worth it relative to the extremely
simple solution here. It feels likely that it would be better to invest
time in improving session sharing than go down that rabbit hole.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>