project: convert camel command names to kebab
Before this, a go command like "ThisCommand" or a python command like
"this_command" would end up in the CLI as a command named thisCommand.

The camel casing always struck me as odd in a CLI interface, so this
changes camel case to kebab, i.e. "this-command".

This is obviously pretty subjective, but it looks a lot better to me and
was very easy to try thanks to the strcase library.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>