revise API: Service is simple getter
Previously you had to pass the same set of params to .service as you
would to .withExec. Now you just call .withExec as before and call
.service to return it as a Service.

The previous approach was toilsome in client use because you had to use
a different struct type with all the same params.

Technically we don't need to change the API at all anymore, so a later
might bring the old API back to maintain compatibility for now.

Signed-off-by: Alex Suraci <alex@dagger.io>