Fix consistency
While these snippets were copied from the Azure Pipelines and AWS CodePipeline guides, I favored making them more consistent with the other CI snippets. We can follow up to fix those guides to be more in line with these, but it requires more changes in the text to do that now.

In general terms, the guides instruct how to setup a local environment, try Dagger locally and try it on CI. But in the cookbook, the snippets assume you already have the needed dependencies commited in git (i.e., next phase), so you install the project's dependencies from the directory, not the packages manually (e.g., `... install dagger-io`).

The most similar instruction for installing dependencies in Python, without installing any more tools in CI, is to let pip install from `pyproject.toml` rather than `-r requirements.txt`. We need a follow-up to fix consistency in installation instructions and guides on how to install the Python SDK.

Note that `npm ci` means `clean install`, it doesn't have anything to do with a `ci` directory. We should probably use the extended version of the command to avoid confusion.

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>