Fix corner case where WithDirectory included wrong contents.
The previous MergeOp optimization missed a corner case where if the
selector of the source dir was the same as the destination path, direct
MergeOp could be triggered and reveal anything above the selector of the
source during the merge.

This fixes it by only triggering direct merge when both src and dest
paths (including any selectors) are "/".

Signed-off-by: Erik Sipsma <erik@sipsma.dev>