Container.import: import to worker content store
Now that we're "under the hood" we can just reach over and import OCI
tarballs directly to the content store!

I kept the caching because the process of importing is still slow even
for a no-op, but now it caches only that part of the import flow. The
idea is we can cache while still being able to deal with content going
away, by caching the manifest and checking that it's still in the
content.Store, and importing again if not.

Signed-off-by: Alex Suraci <alex@dagger.io>