core: support optionally setting explicit mediatypes
We previously made a change to default to using oci mediatypes, which
helped compatibility with certain compression types that are only
supported with oci types.

However, older registries (specifically a user reported Artifactory
versions from pre-2020) don't have OCI support.

This change fixes that by leaving OCI as the default but enabling users
to specify docker mediatypes to be used instead as a fallback.

Signed-off-by: Erik Sipsma <erik@dagger.io>