remove llb.SessionID from Host.Directory for now
This is causing container hostnames to change on every run, causing
caches to bust.

We tried changing Container.WithExec to only set a hostname if the
container exposes ports, but that breaks two use cases:

1. Calling WithExec before WithExposedPort.
1. Using a container as a service without explicitly exposed ports.

We could stomach temporarily breaking the latter case, but the former
case is probably pretty common, since that's what our docs show.

The impact of this change on the other hand should be pretty minor, only
breaking the unannounced `dagger do` command, which is already being
completely rearchitected such that we can probably approach the intended
fix here in a different manner. (#5415)

Signed-off-by: Alex Suraci <alex@dagger.io>