stabilize container hostnames against SessionID (#5474)
* Add test for local dir import cache reuse.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* stabilize container hostnames against SessionID

SessionID is important to keep on host directories, because otherwise
the Buildkit solver's vertex is keyed only on a local client's path,
whose contents may differ when clients are running in containers with
different content mounted to the same path. This primarily affects
setups with a long-running external engine.

To prevent the SessionID from busting caches everywhere, hostnames are
now calculated by taking a stable hash of the *pb.Definition, which is a
pretty invasive process:

First it loops over the inner Ops and removes the relevant ephemeral
metadata from SourceOps. This results in a new Op with a different
digest, so it keeps track of the old and new digests along the way.

Next it loops over the old => new digest mapping and converts any
occurrences of the old digest to the new digest, both in the
definition's OpMetadata and in each inner Op's inputs. Any Ops whose
inputs had to be modified will result in yet another old => new digest
mapping, so this process repeats until everything stabilizes.

Finally, the inner Ops are sorted lexicographically because otherwise
they are in topological order, meaning two root Ops may change places
without changing semantics, but in this case we need everything to be
deterministic.

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix up recursion/looping order

Signed-off-by: Alex Suraci <alex@dagger.io>

* use llbsolver.Load + source policy for stable digests

Signed-off-by: Alex Suraci <alex@dagger.io>

* be more selective with repo content that we pull in

otherwise the git add * call can fail, if there are .gitignored files or
other weird content in your local dir (e.g. bind-mounted repos for
shimming into go.mod replace rules)

Signed-off-by: Alex Suraci <alex@dagger.io>

* Revert "use llbsolver.Load + source policy for stable digests"

works great, but breaks the Darwin build. :(

This reverts commit 7c20b080f96b150139bf3a94a3df161850e307ec.

Signed-off-by: Alex Suraci <alex@dagger.io>

* extract llbsolver.Load from Buildkit

goals: a) replace my hairy yolo'd implementation with a bona-fide one
from upstream, and b) gain memoization along the way.

Signed-off-by: Alex Suraci <alex@dagger.io>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>
Signed-off-by: Alex Suraci <alex@dagger.io>
Co-authored-by: Erik Sipsma <erik@sipsma.dev>