auth: always pass bearer auth related requests to docker auth provider.
Before this, if there was s a matching domain for basic auth creds in the
in-memory credential store, then we error out if one of the methods
related to bearer auth is called.

Now, we instead just passthrough to the docker auth provider. Our
in-mem cred store doesn't support bearer auth tokens, so there's no
reason (I'm aware of) to not just forward the request.

Signed-off-by: Erik Sipsma <erik@dagger.io>