rust: ensure CI generate+test targets run against dev cli+engine.
Before this, the rust CI steps were not actually running w/ the dev CLI
because the SDK doesn't yet support `_EXPERIMENTAL_DAGGER_CLI_BIN`.

This resulted in the steps always downloading the CLI based on the
version last bumped.

Unfortunately, correcting this seems to have possibly revealed a
catch-22 situation in that the generate step fails due to backwards
incompatible changes recently made in our API. Not sure the best way to
go about fixing.

Signed-off-by: Erik Sipsma <erik@dagger.io>