Container: handle pruned import caches
Along the way, change the import tests to use an image built with `apko`
so we're not triple-testing Export and Host APIs. I was getting
different image digests on every run before; by using the result of a
cached container command, they're stable across test runs, allowing me
to buildctl prune out-of-band.

In the future it'd be nice if we had a prune API we could use in tests.

Signed-off-by: Alex Suraci <alex@dagger.io>