feat: add support _EXPERIMENTAL_DAGGER_CLI_BIN to rust
From the root of the rust sdk, run:

    $ env _EXPERIMENTAL_DAGGER_CLI_BIN=../../bin/dagger cargo run --example first-pipeline

The sdk will use dagger binary defined in `_EXPERIMENTAL_DAGGER_CLI_BIN`
to run the pipeline. And modify the engine to try running on local cli
when session port & token doesn't set, if this configuration is not set,
it will fallback to download cli method.

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>