Do not revert mix.exs file after publishing
There should be no need to do this. When this runs in the CI, the
modified file will not be committed. We should not be running this
locally - our CI should publish releases.

When we are testing things and we end up with a modified file, it should
be easy to git revert it. At least that way we will know for sure that
this worked as expected. As it happens, that is exactly what I did
locally when testing the changes in this PR.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>