core: fix max file chunk size
Previously, we were using a max chunk size based on a non-grpc-default
set inside the buildkit client implementation, but the actual grpc
default is much smaller than that.

This didn't impact us previously because we only ever ended up hitting
the codepaths where the larger max message size was used, but after the
recent re-architecture we ended up sometimes using streams where the
smaller grpc default was set.

This updates us to just always assume the 4MB grpc default chunk size is
being enforced, which is safer at the moment than trying to play
whack-a-mole to update the max size on every grpc stream we open. We can
play that game in the future if there's ever a need to tune this for
performance reasons.

Also added integ test coverage for this now.

Signed-off-by: Erik Sipsma <erik@dagger.io>