core: fix setting image config during multiplatform export+publish
We accidentally were exporting the config of the container that export
was being called on, but meant to use the config of the platform variant
in that part of the loop.

Wasn't caught because we didn't have test coverage for image config
being retained for either export or publish. Added that now.

Signed-off-by: Erik Sipsma <erik@dagger.io>