engine: fix combined cache manager to query local cache too.
Before this, we were providing just our remote cache import to the list
of "cms" given to the combined cache manager and setting the "main"
cache to the local cache.

However, it turns out that the combined cache will only query the list
of cache managers provided in "cms"; it will never query the "main"
cache.

Now we provide the local cache both in the list of cms and as the main
cache manager so that it's queried too.

Signed-off-by: Erik Sipsma <erik@dagger.io>