chore: improve elixir bump version (#5527)
* chore: improve elixir bump version

Instead of using VERSION file, replacing version defined in `mix.exs`
and restore it back after published.

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>

* Default Elixir SDK version to 0.0.0

This particular version is specific enough to not be confused with a
genuine release.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Do not revert mix.exs file after publishing

There should be no need to do this. When this runs in the CI, the
modified file will not be committed. We should not be running this
locally - our CI should publish releases.

When we are testing things and we end up with a modified file, it should
be easy to git revert it. At least that way we will know for sure that
this worked as expected. As it happens, that is exactly what I did
locally when testing the changes in this PR.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

---------

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>
Signed-off-by: Gerhard Lazu <gerhard@dagger.io>
Co-authored-by: Gerhard Lazu <gerhard@dagger.io>