We only want to publish the Elixir SDK when there is a tag
We do not want to publish when there is a push to `main`.

This will fix the check which is currently failing on `main`.

FWIW, only the Go SDK does this since it needs to sync to the repository
where users `go get` from, https://github.com/dagger/dagger-go-sdk -
this is when users want to run the latest dev version of Go SDK.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>