fix flaky shell caused by concurrent Read calls
complicated to explain, tl;dr net/http needs to be able to interrupt a
Read, but that's not actually supported by grpchijack. previous attempts
allowed the Read call to unblock, but it still left the concurrent call
running.

now we start a single dedicated Read goroutine that writes to a
bytes.Buffer, which we concurrently read from, ensuring there's only one
actual Reader, and allowing call sites to unblock while not dropping
data.

Signed-off-by: Alex Suraci <alex@dagger.io>