core: hide merge vertexes from output
The update to use merge-op to optimize copies also resulted in us
receiving tons of progress vertexes about merges. These are kind of
confusing to users since merge is entirely an internal optimization and
the format we receive them in can be extremely ugly due to the fact that
they don't show the flattened merge that actually takes place.

Now that vertex is marked as internal and thus shouldn't show up in
output by default.

Signed-off-by: Erik Sipsma <erik@dagger.io>