Temporarily suspend auto-triggering the Elixir SDK workflow
Tests started failing intermittently since the fix in
https://github.com/dagger/dagger/pull/5628 and it's not clear why. FTR:

    ```diff
    diff --git a/sdk/elixir/lib/dagger/engine_conn.ex b/sdk/elixir/lib/dagger/engine_conn.ex
    index 9b4ed912..acd97199 100644
    --- a/sdk/elixir/lib/dagger/engine_conn.ex
    +++ b/sdk/elixir/lib/dagger/engine_conn.ex
    @@ -16,7 +16,8 @@ defmodule Dagger.EngineConn do
           _otherwise ->
             case from_local_cli(opts) do
               {:ok, conn} -> {:ok, conn}
    -          _otherwise -> from_remote_cli(opts)
    +          {:error, :no_executable} -> from_remote_cli(opts)
    +          {:error, :session_timeout} -> {:error, :session_timeout}
             end
         end
       end
    @@ -41,7 +42,6 @@ defmodule Dagger.EngineConn do
           start_cli_session(bin_path, opts)
         else
           nil -> {:error, :no_executable}
    -      otherwise -> otherwise
         end
       end
    ```

Tried a few local changes, and the final result is that the connection
to the Engine keeps timing out in 80% of the test runs. We want to
produce a release and cannot do so with failing checks.

Having hit the timebox, the easiest thing is to disable this for now and
figure out how to improve it later.

I have also reverted the Elixir SDK client timeout from 30s to 10s since
the fix was clearly not helping (if anything, it's making the failures
happen later).

Part of this extracted a _hack_make reusable workflow that we can keep
referencing instead of repeating ourselves. This is a temporary measure,
we want to drop all this YAML in favour of code. Btw, notice the
improved setup-go step caching config.

Another small step that we took in this new GitHub workflow is to:
- use our own runners when running in the context of dagger/dagger
- use our paid runners when running in the context of PRs
- otherwise use free GitHub runners

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>